fix(account.contacts): contact change request redirection (#4463)
ref: DTRSD-29941
