+++ packages/manager/apps/dedicated/client/app/account/contacts/index.js
@@ -1,4 +1,7 @@
 import angular from 'angular';
+import get from 'lodash/get';
+import has from 'lodash/has';
+import last from 'lodash/last';
 import '@uirouter/angularjs';
 import 'oclazyload';
 
@@ -20,6 +23,29 @@ angular
           ).then((mod) => $ocLazyLoad.inject(mod.default || mod));
         },
       });
+    },
+  )
+  .config(
+    /* @ngInject */ ($urlRouterProvider) => {
+      // ensure compatibility with links sended by emails
+      // like #/useraccount/contacts/1124580?tab=REQUESTS&token=myToken
+      // make a redirect to the new url of ui route
+      $urlRouterProvider.when(
+        /^\/useraccount\/contacts\/[0-9]+$/,
+        ($location, $state) => {
+          const hasToken = has($location.search(), 'token');
+          const requestTabAsked = get($location.search(), 'tab') === 'REQUESTS';
+
+          if (!hasToken || !requestTabAsked) {
+            return false;
+          }
+
+          const taskId = last($location.path().split('/'));
+          const token = get($location.search(), 'token');
+
+          return $state.go('app.account.contacts.requests', { taskId, token });
+        },
+      );
     },
   )
   .service('UserAccountContactsService', service);
